LayoutTests/media/media-source/media-source-real-end-of-stream-readyState.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-end-of-stream-readyState</title>
    <script src="mp4-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    // Real-MP4 equivalent of media-source-end-of-stream-readyState.html
    // Tests that endOfStream() normalizes duration to the end of buffered data.

    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        source = new MediaSource();
        run('video.src = URL.createObjectURL(source)');
        await waitFor(source, 'sourceopen');

        run(`sourceBuffer = source.addSourceBuffer('${MP4.VIDEO_TYPE}')`);

        // Init with duration hint of 10 seconds
        var initData = MP4.initSegment({
            timescale: 1000,
            tracks: [{ id: 1, type: 'video', timescale: 1000 }]
        });
        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        // Set explicit duration to 10
        run('source.duration = 10');

        // Append 5 seconds of data
        var media = MP4.mediaSegment({
            sequenceNumber: 1,
            tracks: [{ id: 1, type: 'video', baseDecodeTime: 0, samples: [
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
            ]}]
        });
        sourceBuffer.appendBuffer(media);
        await waitFor(sourceBuffer, 'updateend');

        testExpected('source.duration.toFixed(3)', 10);
        testExpected('sourceBuffer.buffered.end(0).toFixed(3)', 5);

        // endOfStream should normalize duration to end of buffered data
        run('source.endOfStream()');
        testExpected('source.duration.toFixed(3)', 5);
        testExpected('sourceBuffer.updating', false);

        endTest();
    }
    </script>
</head>
<body onload="runTest()">
    <video></video>
</body>
</html>
