LayoutTests/media/media-source/media-source-real-end-of-stream-readyState-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer('video/mp4; codecs="avc1.4d401e"'))
EVENT(updateend)
RUN(source.duration = 10)
EVENT(updateend)
EXPECTED (source.duration.toFixed(3) == '10') OK
EXPECTED (sourceBuffer.buffered.end(0).toFixed(3) == '5') OK
RUN(source.endOfStream())
EXPECTED (source.duration.toFixed(3) == '5') OK
EXPECTED (sourceBuffer.updating == 'false') OK
END OF TEST
